Commenting is done properly

+++ PolarComputerCycleAnalysisTests/SummaryTests.cs
@@ -8,6 +8,9 @@ using System.Threading.Tasks;
 
 namespace PolarComputerCycleAnalysis.Tests
 {
+    /// <summary>
+    /// testing whether this method finds the maximum value from arraylist or not
+    /// </summary>
     [TestClass()]
     public class SummaryTests
     {
@@ -18,27 +21,36 @@ namespace PolarComputerCycleAnalysis.Tests
             Assert.AreEqual(18, maxValue);
         }
 
+        /// <summary>
+        /// testing whether this method finds the manimum value from arraylist or not
+        /// </summary>
         [TestMethod()]
         public void FindMinTest()
         {
             int val = Summary.FindMin(new List<string> { "15", "10", "4", "18", "16" });
             Assert.AreEqual(4, val);
         }
-
+        /// <summary>
+        /// testing whether this method finds the average value from arraylist or not
+        /// </summary>
         [TestMethod()]
         public void FindAverageTest()
         {
             double val = Summary.FindAverage(new List<string> { "15", "10", "4", "18", "16" });
             Assert.AreEqual(12, val);
         }
-
+        /// <summary>
+        /// testing whether this method returns the total sum value from arraylist or not
+        /// </summary>
         [TestMethod()]
         public void FindSumTest()
         {
-            int val = Summary.FindSum(new List<string> { "15", "10", "4", "18", "16" });
+            double val = Summary.FindSum(new List<string> { "15", "10", "4", "18", "16" });
             Assert.AreEqual(63, val);
         }
-
+        /// <summary>
+        /// testing whether this method converts string value of date into correct format of date or not
+        /// </summary>
         [TestMethod()]
         public void ConvertToDate()
         {
